Compute the Automorphisms of Mutational Events Between two Codon Sequences Represented in a Given Abelian group.

Given two codon sequences represented in a given Abelian group, this function computes the automorphisms describing codon mutational events. Basically, this function is a wrapping to call the corresponding function for a specified Abelian group.

automorphism(seqs = NULL, filepath = NULL, group = "Z4", ...)

# S4 method for DNAStringSet_OR_NULL
automorphism(
  seqs = NULL,
  filepath = NULL,
  group = c("Z5", "Z64", "Z125", "Z5^3"),
  cube = c("ACGT", "TGCA"),
  cube_alt = c("CATG", "GTAC"),
  nms = NULL,
  start = NA,
  end = NA,
  chr = 1L,
  strand = "+"
)

Arguments

filepath
A character vector containing the path to a file in fasta format to be read. This argument must be given if codon & base arguments are not provided.

group
A character string denoting the group representation for the given base or codon as shown in reference (1).

cube, cube_alt
A character string denoting pairs of the 24 Genetic-code cubes, as given in references (2-3). That is, the base pairs from the given cubes must be complementary each other. Such a cube pair are call \(dual cubes\) and, as shown in reference (3), each pair integrates group.

nms
Optional. Only used if the DNA sequence alignment provided carries more than two sequences. A character string giving short names for the alignments to be compared. If not given then the automorphisms between pairwise alignment are named as: "aln_1", "aln_2", and so on.

start, end, chr, strand
Optional parameters required to build a GRanges-class. If not provided the default values given for the function definition will be used.

seq
An object from a DNAStringSet or DNAMultipleAlignment class carrying the DNA pairwise alignment of two sequences. The pairwise alignment provided in argument seq or the 'fasta' file filepath must correspond to codon sequences.

Value

This function returns a Automorphism-class object with four columns on its metacolumn named: seq1, seq2, autm, and cube.

Methods

automorphismByRanges:

This function returns a GRanges-class object. Consecutive mutational events (on the codon sequence) described by automorphisms on a same cube are grouped in a range.

automorphismByCoef

This function returns a GRanges-class object. Consecutive mutational events (on the codon sequence) described by the same automorphism coefficients are grouped in a range.

getAutomorphisms

This function returns an AutomorphismList-class object as a list of Automorphism-class objects, which inherits from GRanges-class objects.

conserved_regions

Returns a AutomorphismByCoef class object containing the requested regions.

Details

Herein, automorphisms are algebraic descriptions of mutational event observed in codon sequences represented on different Abelian groups. In particular, as described in references (3-4), for each representation of the codon set on a defined Abelian group there are 24 possible isomorphic Abelian groups. These Abelian groups can be labeled based on the DNA base-order used to generate them. The set of 24 Abelian groups can be described as a group isomorphic to the symmetric group of degree four (\(S_4\), see reference (4)). Function automorphismByRanges permits the classification of the pairwise alignment of protein-coding sub-regions based on the mutational events observed on it and on the genetic-code cubes that describe them.

Automorphisms in Z5, Z64 and Z125 are described as functions \(f(x) = k x mod 64\) and \(f(x) = k x mod 125\), where k and x are elements from the set of integers modulo 64 or modulo 125, respectively. If an automorphism cannot be found on any of the cubes provided in the argument \(cube\), then function automorphism will search for automorphisms in the cubes provided in the argument \(cube_alt\).

Automorphisms in Z5^3' are described as functions \(f(x) = Ax mod Z5\), where A is diagonal matrix.

Arguments cube and cube_alt must be pairs of' dual cubes (see section 2.4 from reference 4).

References

Sanchez R, Morgado E, Grau R. Gene algebra from a genetic code algebraic structure. J Math Biol. 2005 Oct;51(4):431-57. doi: 10.1007/s00285-005-0332-8. Epub 2005 Jul 13. PMID: 16012800. ( PDF).

Robersy Sanchez, Jesús Barreto (2021) Genomic Abelian Finite Groups. doi: 10.1101/2021.06.01.446543.

M. V José, E.R. Morgado, R. Sánchez, T. Govezensky, The 24 possible algebraic representations of the standard genetic code in six or in three dimensions, Adv. Stud. Biol. 4 (2012) 119–152.PDF.

R. Sanchez. Symmetric Group of the Genetic–Code Cubes. Effect of the Genetic–Code Architecture on the Evolutionary Process MATCH Commun. Math. Comput. Chem. 79 (2018) 527-560. PDF

See also

Examples

## Load a pairwise alignment
data(aln)
aln

#> DNAStringSet object of length 2:
#>     width seq
#> [1]    51 ACCTATGTTGGTATT---GCGCTCCAACTCCTTGGCTCTAGCTCACTACAT
#> [2]    51 ATCTATGTTGGTATTACGACGCTCCAATTCCTTGGGTCC------CTCCTT

## Automorphism on "Z5^3"
autms <- automorphism(seq = aln, group = "Z5^3")

#> Error in validObject(.Object): invalid class “CodonGroup” object: 
#>     *** Argument 'x' is not a CodonGroup-class object.The slot 'group' is
#>     not present or wrong naming.

autms

#> Automorphism object with 7100 ranges and 6 metadata columns:
#>          seqnames    ranges strand |        seq1        seq2    coord1
#>             <Rle> <IRanges>  <Rle> | <character> <character> <numeric>
#>      [1]        1         1      + |         ATG         ATG        50
#>      [2]        1         2      + |         GAG         GAG        10
#>      [3]        1         3      + |         AGC         AGC        33
#>      [4]        1         4      + |         CTT         CTT        55
#>      [5]        1         5      + |         GTT         GTC        59
#>      ...      ...       ...    ... .         ...         ...       ...
#>   [7096]        1      7096      + |         CTT         CTT        55
#>   [7097]        1      7097      + |         GTT         GTT        59
#>   [7098]        1      7098      + |         AAT         AAC         3
#>   [7099]        1      7099      + |         AAC         AAC         1
#>   [7100]        1      7100      + |         TAA         TAA        12
#>             coord2      autm        cube
#>          <numeric> <numeric> <character>
#>      [1]        50         1        ACGT
#>      [2]        10         1        ACGT
#>      [3]        33         1        ACGT
#>      [4]        55         1        ACGT
#>      [5]        57        27        ACGT
#>      ...       ...       ...         ...
#>   [7096]        55         1        ACGT
#>   [7097]        59         1        ACGT
#>   [7098]         1        43        ACGT
#>   [7099]         1         1        ACGT
#>   [7100]        12         1        ACGT
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths

## Automorphism on "Z64"
autms <- automorphism(seq = aln, group = "Z64")
autms

#> Automorphism object with 17 ranges and 6 metadata columns:
#>        seqnames    ranges strand |        seq1        seq2    coord1    coord2
#>           <Rle> <IRanges>  <Rle> | <character> <character> <numeric> <numeric>
#>    [1]        1         1      + |         ACC         ATC        17        49
#>    [2]        1         2      + |         TAT         TAT        15        15
#>    [3]        1         3      + |         GTT         GTT        59        59
#>    [4]        1         4      + |         GGT         GGT        43        43
#>    [5]        1         5      + |         ATT         ATT        51        51
#>    ...      ...       ...    ... .         ...         ...       ...       ...
#>   [13]        1        13      + |         TCT         TCC        31        29
#>   [14]        1        14      + |         AGC         ---        52        NA
#>   [15]        1        15      + |         TCA         ---         9        NA
#>   [16]        1        16      - |         CTA         CTC        52        53
#>   [17]        1        17      + |         CAT         CTT         7        55
#>             autm        cube
#>        <numeric> <character>
#>    [1]        33        ACGT
#>    [2]         1        ACGT
#>    [3]         1        ACGT
#>    [4]         1        ACGT
#>    [5]         1        ACGT
#>    ...       ...         ...
#>   [13]         3        ACGT
#>   [14]        NA        None
#>   [15]        NA        None
#>   [16]        30        TGCA
#>   [17]        17        ACGT
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths

## Grouping into ranges the automorphisms by cubes 
automorphismByRanges(autms)

#> GRanges object with 6 ranges and 1 metadata column:
#>       seqnames    ranges strand |        cube
#>          <Rle> <IRanges>  <Rle> | <character>
#>   [1]        1       1-5      + |        ACGT
#>   [2]        1         6      + |        None
#>   [3]        1      7-13      + |        ACGT
#>   [4]        1     14-15      + |        None
#>   [5]        1        16      - |        TGCA
#>   [6]        1        17      + |        ACGT
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths

## Automorphism on "Z64" from position 1 to 33
autms <- automorphism(seq = aln,
                      group = "Z64",
                      start = 1,
                      end = 33)
autms

#> Automorphism object with 11 ranges and 6 metadata columns:
#>        seqnames    ranges strand |        seq1        seq2    coord1    coord2
#>           <Rle> <IRanges>  <Rle> | <character> <character> <numeric> <numeric>
#>    [1]        1         1      + |         ACC         ATC        17        49
#>    [2]        1         2      + |         TAT         TAT        15        15
#>    [3]        1         3      + |         GTT         GTT        59        59
#>    [4]        1         4      + |         GGT         GGT        43        43
#>    [5]        1         5      + |         ATT         ATT        51        51
#>    [6]        1         6      + |         ---         ACG        NA         7
#>    [7]        1         7      + |         GCG         ACG        26        18
#>    [8]        1         8      + |         CTC         CTC        53        53
#>    [9]        1         9      + |         CAA         CAA         4         4
#>   [10]        1        10      + |         CTC         TTC        53        61
#>   [11]        1        11      + |         CTT         CTT        55        55
#>             autm        cube
#>        <numeric> <character>
#>    [1]        33        ACGT
#>    [2]         1        ACGT
#>    [3]         1        ACGT
#>    [4]         1        ACGT
#>    [5]         1        ACGT
#>    [6]        NA        None
#>    [7]        45        ACGT
#>    [8]         1        ACGT
#>    [9]         1        ACGT
#>   [10]        41        ACGT
#>   [11]         1        ACGT
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths

## Grouping into ranges the automorphisms by cubes 
automorphismByRanges(autms)

#> GRanges object with 3 ranges and 1 metadata column:
#>       seqnames    ranges strand |        cube
#>          <Rle> <IRanges>  <Rle> | <character>
#>   [1]        1       1-5      + |        ACGT
#>   [2]        1         6      + |        None
#>   [3]        1      7-11      + |        ACGT
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths