Methods for Automorphism-class and AutomorphismList-class Objects

Several methods are available to be applied on Automorphism-class and AutomorphismList-class objects.

Replace the element names in AutomorphismList-class object

Usage

as.AutomorphismList(x, grs = GRanges(), ...)

# S4 method for GRangesList,GRanges_OR_NULL
as.AutomorphismList(x, grs = GRanges(), ...)

# S4 method for list,GRanges_OR_NULL
as.AutomorphismList(x, grs = GRanges(), ...)

# S4 method for AutomorphismList
names(x)

# S4 method for AutomorphismList,ANY
names(x) <- value

# S4 method for AutomorphismList
as.list(x)

# S4 method for AutomorphismList
show(object)

Arguments

x

A DataFrame or a automorphisms class object.

object

An object from AutomorphismList-class.

gr

A GRanges-class object.

Value

The returned value depends on (and it is understood) from the method called.

Automorphism-class methods

as(from, "Automorphism"):

Permits the transformation of a data.frame or a DataFrame-class object into Automorphism-class object if the proper columns are provided.

Methods from GRanges-class can also be applied

AutomorphismList-class methods

as.AutomorphismList(x):

as.AutomorphismList function transform a list of GRanges-class, a GRangesList-class, a list of data.frame or a DataFrame-class objects into a AutomorphismList-class object.

unlist(x)

It transforms a AutomorphismList-class object into an Automorphism-class object.

as.list(x)

It transforms a list of Automorphism-class objects into an AutomorphismList-class object.

as(x, "GRangesList")

It transforms a 'GRangesList' of Automorphism-class objects into an AutomorphismList-class object.

names(x)

To get the element's names from an AutomorphismList-class object.

names(x) <- value

To assign names to the element from an AutomorphismList-class object.

AutomorphismByCoefList-class methods

unlist(x):

It transforms a AutomorphismByCoefList-class object into an AutomorphismByCoef-class object.

as(x, "AutomorphismByCoefList")

It transforms a 'list' of AutomorphismByCoef-class object into an AutomorphismByCoefList-class object.

See also

Examples

## Load datasets
data(autm, brca1_autm)

## Transforming a list of Automorphisms into an AutomorphismList object
lista <- list(human = brca1_autm[[1]], gorilla = brca1_autm[[2]])
as.AutomorphismList(lista)
#> AutomorphismList object of length: 2
#> names(2): human gorilla 
#> ------- 
#> Automorphism object with 761 ranges and 6 metadata columns:
#>         seqnames    ranges strand |        seq1        seq2    coord1    coord2
#>            <Rle> <IRanges>  <Rle> | <character> <character> <numeric> <numeric>
#>     [1]        1         1      + |         ATG         ATG        50        50
#>     [2]        1         2      + |         GAT         GAT        11        11
#>     [3]        1         3      + |         TTA         TTA        60        60
#>     [4]        1         4      + |         TCT         TCT        31        31
#>     [5]        1         5      + |         GCT         GCT        27        27
#>     ...      ...       ...    ... .         ...         ...       ...       ...
#>   [757]        1       757      + |         CAC         CAC         5         5
#>   [758]        1       758      + |         AGC         AGC        33        33
#>   [759]        1       759      + |         CAC         CAC         5         5
#>   [760]        1       760      + |         TAC         TAC        13        13
#>   [761]        1       761      + |         TGA         TGA        44        44
#>              autm        cube
#>         <numeric> <character>
#>     [1]         1        ACGT
#>     [2]         1        ACGT
#>     [3]         1        ACGT
#>     [4]         1        ACGT
#>     [5]         1        ACGT
#>     ...       ...         ...
#>   [757]         1        ACGT
#>   [758]         1        ACGT
#>   [759]         1        ACGT
#>   [760]         1        ACGT
#>   [761]         1        ACGT
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths
#> ...
#> <1 more DFrame element(s)>
#> Two slots: 'DataList' & 'SeqRanges'
#> ------- 

## Alternatively we can set
aut <- as.list(brca1_autm[1:2])
class(aut)
#> [1] "list"

## And reverse it
aut <- as.AutomorphismList(aut)
aut
#> AutomorphismList object of length: 2
#> names(2): human_1.human_2 human_1.gorilla_1 
#> ------- 
#> Automorphism object with 761 ranges and 6 metadata columns:
#>         seqnames    ranges strand |        seq1        seq2    coord1    coord2
#>            <Rle> <IRanges>  <Rle> | <character> <character> <numeric> <numeric>
#>     [1]        1         1      + |         ATG         ATG        50        50
#>     [2]        1         2      + |         GAT         GAT        11        11
#>     [3]        1         3      + |         TTA         TTA        60        60
#>     [4]        1         4      + |         TCT         TCT        31        31
#>     [5]        1         5      + |         GCT         GCT        27        27
#>     ...      ...       ...    ... .         ...         ...       ...       ...
#>   [757]        1       757      + |         CAC         CAC         5         5
#>   [758]        1       758      + |         AGC         AGC        33        33
#>   [759]        1       759      + |         CAC         CAC         5         5
#>   [760]        1       760      + |         TAC         TAC        13        13
#>   [761]        1       761      + |         TGA         TGA        44        44
#>              autm        cube
#>         <numeric> <character>
#>     [1]         1        ACGT
#>     [2]         1        ACGT
#>     [3]         1        ACGT
#>     [4]         1        ACGT
#>     [5]         1        ACGT
#>     ...       ...         ...
#>   [757]         1        ACGT
#>   [758]         1        ACGT
#>   [759]         1        ACGT
#>   [760]         1        ACGT
#>   [761]         1        ACGT
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths
#> ...
#> <1 more DFrame element(s)>
#> Two slots: 'DataList' & 'SeqRanges'
#> ------- 

## Let's get the element names from object 'aut'
names(aut)
#> [1] "human_1.human_2"   "human_1.gorilla_1"

## Let's assign new names
names(aut) <- c("human_1", "gorilla_1")
names(aut)
#> [1] "human_1"   "gorilla_1"

## Transforming a GRangesList of Automorphisms into an AutomorphismList
## object
lista <- as(lista, "GRangesList")
as.AutomorphismList(lista)
#> AutomorphismList object of length: 2
#> names(2): human gorilla 
#> ------- 
#> Automorphism object with 761 ranges and 6 metadata columns:
#>         seqnames    ranges strand |        seq1        seq2    coord1    coord2
#>            <Rle> <IRanges>  <Rle> | <character> <character> <numeric> <numeric>
#>     [1]        1         1      + |         ATG         ATG        50        50
#>     [2]        1         2      + |         GAT         GAT        11        11
#>     [3]        1         3      + |         TTA         TTA        60        60
#>     [4]        1         4      + |         TCT         TCT        31        31
#>     [5]        1         5      + |         GCT         GCT        27        27
#>     ...      ...       ...    ... .         ...         ...       ...       ...
#>   [757]        1       757      + |         CAC         CAC         5         5
#>   [758]        1       758      + |         AGC         AGC        33        33
#>   [759]        1       759      + |         CAC         CAC         5         5
#>   [760]        1       760      + |         TAC         TAC        13        13
#>   [761]        1       761      + |         TGA         TGA        44        44
#>              autm        cube
#>         <numeric> <character>
#>     [1]         1        ACGT
#>     [2]         1        ACGT
#>     [3]         1        ACGT
#>     [4]         1        ACGT
#>     [5]         1        ACGT
#>     ...       ...         ...
#>   [757]         1        ACGT
#>   [758]         1        ACGT
#>   [759]         1        ACGT
#>   [760]         1        ACGT
#>   [761]         1        ACGT
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths
#> ...
#> <1 more DFrame element(s)>
#> Two slots: 'DataList' & 'SeqRanges'
#> ------- 

## Transform a AutomorphismList-class object into an Automorphism-class
## object 
unlist(brca1_autm[1:2])
#> Automorphism object with 1522 ranges and 6 metadata columns:
#>                     seqnames    ranges strand |        seq1        seq2
#>                        <Rle> <IRanges>  <Rle> | <character> <character>
#>     human_1.human_2        1         1      + |         ATG         ATG
#>     human_1.human_2        1         2      + |         GAT         GAT
#>     human_1.human_2        1         3      + |         TTA         TTA
#>     human_1.human_2        1         4      + |         TCT         TCT
#>     human_1.human_2        1         5      + |         GCT         GCT
#>                 ...      ...       ...    ... .         ...         ...
#>   human_1.gorilla_1        1       757      + |         CAC         CAC
#>   human_1.gorilla_1        1       758      + |         AGC         AGC
#>   human_1.gorilla_1        1       759      + |         CAC         CAC
#>   human_1.gorilla_1        1       760      + |         TAC         TAC
#>   human_1.gorilla_1        1       761      + |         TGA         TGA
#>                        coord1    coord2      autm        cube
#>                     <numeric> <numeric> <numeric> <character>
#>     human_1.human_2        50        50         1        ACGT
#>     human_1.human_2        11        11         1        ACGT
#>     human_1.human_2        60        60         1        ACGT
#>     human_1.human_2        31        31         1        ACGT
#>     human_1.human_2        27        27         1        ACGT
#>                 ...       ...       ...       ...         ...
#>   human_1.gorilla_1         5         5         1        ACGT
#>   human_1.gorilla_1        33        33         1        ACGT
#>   human_1.gorilla_1         5         5         1        ACGT
#>   human_1.gorilla_1        13        13         1        ACGT
#>   human_1.gorilla_1        44        44         1        ACGT
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths

## Let's transform a AutomorphismByCoefList-class object into an 
## AutomorphismByCoef-class object
data("autby_coef")
unlist(autby_coef[1:2])
#> AutomorphismByCoef object with 683 ranges and 5 metadata columns:
#>                     seqnames    ranges strand |        seq1        seq2
#>                        <Rle> <IRanges>  <Rle> | <character> <character>
#>     human_1.human_2        1     1-238      + |         ATG         ATG
#>     human_1.human_2        1     1-238      + |         GAT         GAT
#>     human_1.human_2        1     1-238      + |         TTA         TTA
#>     human_1.human_2        1     1-238      + |         TCT         TCT
#>     human_1.human_2        1     1-238      + |         GCT         GCT
#>                 ...      ...       ...    ... .         ...         ...
#>   human_1.gorilla_1        1   729-761      + |         CCC         CCC
#>   human_1.gorilla_1        1   729-761      + |         ATC         ATC
#>   human_1.gorilla_1        1   729-761      + |         CAC         CAC
#>   human_1.gorilla_1        1   729-761      + |         AGC         AGC
#>   human_1.gorilla_1        1   729-761      + |         TGA         TGA
#>                          autm    mut_type        cube
#>                     <numeric> <character> <character>
#>     human_1.human_2         1         HHH        ACGT
#>     human_1.human_2         1         HHH        ACGT
#>     human_1.human_2         1         HHH        ACGT
#>     human_1.human_2         1         HHH        ACGT
#>     human_1.human_2         1         HHH        ACGT
#>                 ...       ...         ...         ...
#>   human_1.gorilla_1         1         HHH        ACGT
#>   human_1.gorilla_1         1         HHH        ACGT
#>   human_1.gorilla_1         1         HHH        ACGT
#>   human_1.gorilla_1         1         HHH        ACGT
#>   human_1.gorilla_1         1         HHH        ACGT
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths

## Herein a 'list' object of AutomorphismByCoef-class objects
lista <- list(human = autby_coef[[1]], gorilla = autby_coef[[2]])

## Let's transform the the last list 'lista' into an
## AutomorphismByCoefList-class object
aut <- as(lista, "AutomorphismByCoefList")
aut
#> AutomorphismByCoefList object of length 2:
#> $human
#> AutomorphismByCoef object with 239 ranges and 5 metadata columns:
#>         seqnames    ranges strand |        seq1        seq2      autm
#>            <Rle> <IRanges>  <Rle> | <character> <character> <numeric>
#>     [1]        1     1-238      + |         ATG         ATG         1
#>     [2]        1     1-238      + |         GAT         GAT         1
#>     [3]        1     1-238      + |         TTA         TTA         1
#>     [4]        1     1-238      + |         TCT         TCT         1
#>     [5]        1     1-238      + |         GCT         GCT         1
#>     ...      ...       ...    ... .         ...         ...       ...
#>   [235]        1   511-761      + |         CCC         CCC         1
#>   [236]        1   511-761      + |         CTT         CTT         1
#>   [237]        1   511-761      + |         CCT         CCT         1
#>   [238]        1   511-761      + |         ATA         ATA         1
#>   [239]        1   511-761      + |         TGA         TGA         1
#>            mut_type        cube
#>         <character> <character>
#>     [1]         HHH        ACGT
#>     [2]         HHH        ACGT
#>     [3]         HHH        ACGT
#>     [4]         HHH        ACGT
#>     [5]         HHH        ACGT
#>     ...         ...         ...
#>   [235]         HHH        ACGT
#>   [236]         HHH        ACGT
#>   [237]         HHH        ACGT
#>   [238]         HHH        ACGT
#>   [239]         HHH        ACGT
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths
#> 
#> $gorilla
#> AutomorphismByCoef object with 444 ranges and 5 metadata columns:
#>         seqnames    ranges strand |        seq1        seq2      autm
#>            <Rle> <IRanges>  <Rle> | <character> <character> <numeric>
#>     [1]        1      1-59      + |         ATG         ATG         1
#>     [2]        1      1-59      + |         GAT         GAT         1
#>     [3]        1      1-59      + |         TTA         TTA         1
#>     [4]        1      1-59      + |         TCT         TCT         1
#>     [5]        1      1-59      + |         GCT         GCT         1
#>     ...      ...       ...    ... .         ...         ...       ...
#>   [440]        1   729-761      + |         CCC         CCC         1
#>   [441]        1   729-761      + |         ATC         ATC         1
#>   [442]        1   729-761      + |         CAC         CAC         1
#>   [443]        1   729-761      + |         AGC         AGC         1
#>   [444]        1   729-761      + |         TGA         TGA         1
#>            mut_type        cube
#>         <character> <character>
#>     [1]         HHH        ACGT
#>     [2]         HHH        ACGT
#>     [3]         HHH        ACGT
#>     [4]         HHH        ACGT
#>     [5]         HHH        ACGT
#>     ...         ...         ...
#>   [440]         HHH        ACGT
#>   [441]         HHH        ACGT
#>   [442]         HHH        ACGT
#>   [443]         HHH        ACGT
#>   [444]         HHH        ACGT
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths
#> 

## Let's get the element names from object 'aut'
names(aut)
#> [1] "human"   "gorilla"

## Let's assign new names
names(aut) <- c("human_1", "gorilla_1")
names(aut)
#> [1] "human_1"   "gorilla_1"
## Load a DNA sequence alignment
data("brca1_autm", package = "GenomAutomorphism")

## The list of the first three elements
autm_list <- as.list(brca1_autm[1:3])
autm_list
#> $human_1.human_2
#> Automorphism object with 761 ranges and 6 metadata columns:
#>         seqnames    ranges strand |        seq1        seq2    coord1    coord2
#>            <Rle> <IRanges>  <Rle> | <character> <character> <numeric> <numeric>
#>     [1]        1         1      + |         ATG         ATG        50        50
#>     [2]        1         2      + |         GAT         GAT        11        11
#>     [3]        1         3      + |         TTA         TTA        60        60
#>     [4]        1         4      + |         TCT         TCT        31        31
#>     [5]        1         5      + |         GCT         GCT        27        27
#>     ...      ...       ...    ... .         ...         ...       ...       ...
#>   [757]        1       757      + |         CAC         CAC         5         5
#>   [758]        1       758      + |         AGC         AGC        33        33
#>   [759]        1       759      + |         CAC         CAC         5         5
#>   [760]        1       760      + |         TAC         TAC        13        13
#>   [761]        1       761      + |         TGA         TGA        44        44
#>              autm        cube
#>         <numeric> <character>
#>     [1]         1        ACGT
#>     [2]         1        ACGT
#>     [3]         1        ACGT
#>     [4]         1        ACGT
#>     [5]         1        ACGT
#>     ...       ...         ...
#>   [757]         1        ACGT
#>   [758]         1        ACGT
#>   [759]         1        ACGT
#>   [760]         1        ACGT
#>   [761]         1        ACGT
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths
#> 
#> $human_1.gorilla_1
#> Automorphism object with 761 ranges and 6 metadata columns:
#>         seqnames    ranges strand |        seq1        seq2    coord1    coord2
#>            <Rle> <IRanges>  <Rle> | <character> <character> <numeric> <numeric>
#>     [1]        1         1      + |         ATG         ATG        50        50
#>     [2]        1         2      + |         GAT         GAT        11        11
#>     [3]        1         3      + |         TTA         TTA        60        60
#>     [4]        1         4      + |         TCT         TCT        31        31
#>     [5]        1         5      + |         GCT         GCT        27        27
#>     ...      ...       ...    ... .         ...         ...       ...       ...
#>   [757]        1       757      + |         CAC         CAC         5         5
#>   [758]        1       758      + |         AGC         AGC        33        33
#>   [759]        1       759      + |         CAC         CAC         5         5
#>   [760]        1       760      + |         TAC         TAC        13        13
#>   [761]        1       761      + |         TGA         TGA        44        44
#>              autm        cube
#>         <numeric> <character>
#>     [1]         1        ACGT
#>     [2]         1        ACGT
#>     [3]         1        ACGT
#>     [4]         1        ACGT
#>     [5]         1        ACGT
#>     ...       ...         ...
#>   [757]         1        ACGT
#>   [758]         1        ACGT
#>   [759]         1        ACGT
#>   [760]         1        ACGT
#>   [761]         1        ACGT
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths
#> 
#> $human_1.gorilla_2
#> Automorphism object with 761 ranges and 6 metadata columns:
#>         seqnames    ranges strand |        seq1        seq2    coord1    coord2
#>            <Rle> <IRanges>  <Rle> | <character> <character> <numeric> <numeric>
#>     [1]        1         1      + |         ATG         ATG        50        50
#>     [2]        1         2      + |         GAT         GAT        11        11
#>     [3]        1         3      + |         TTA         TTA        60        60
#>     [4]        1         4      + |         TCT         TCT        31        31
#>     [5]        1         5      + |         GCT         GCT        27        27
#>     ...      ...       ...    ... .         ...         ...       ...       ...
#>   [757]        1       757      + |         CAC         CAC         5         5
#>   [758]        1       758      + |         AGC         AGC        33        33
#>   [759]        1       759      + |         CAC         CAC         5         5
#>   [760]        1       760      + |         TAC         TAC        13        13
#>   [761]        1       761      + |         TGA         TGA        44        44
#>              autm        cube
#>         <numeric> <character>
#>     [1]         1        ACGT
#>     [2]         1        ACGT
#>     [3]         1        ACGT
#>     [4]         1        ACGT
#>     [5]         1        ACGT
#>     ...       ...         ...
#>   [757]         1        ACGT
#>   [758]         1        ACGT
#>   [759]         1        ACGT
#>   [760]         1        ACGT
#>   [761]         1        ACGT
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths
#>