Definition of GRangesMatrixSeq-class

This is a very simple flexible class to store DNA and aminoacid aligned sequences together with their physicochemical properties. That is, a place where each aminoacid or codon from the sequence is represented by numerical value from a physicochemical index.

Constructor for 'GRangesMatrixSeq-class' object.

Usage

GRangesMatrixSeq(
  object = NULL,
  seqnames = Rle(factor()),
  start = integer(0),
  end = integer(0),
  ranges = IRanges(),
  strands = Rle(strand()),
  elementMetadata = DataFrame(),
  seqinfo = NULL,
  seqs = character(),
  names = character(),
  aaindex = character(),
  phychem = character(),
  accession = character()
)

Arguments

object

If provided, it must be a GRangesMatrixSeq-class object and in this case

seqnames, start, end, ranges, strand, elementMetadata, seqinfo

The same as in GRanges

seqs, names, aaindex, phychem, accession

The same as in MatrixSeq.

Value

Given the slot values, it defines a MatrixList-class.

Only used to specify signature in the S4 setMethod.

Only used to specify signature in the S4 setMethod.

Details

This is a convenient function to transform a MatrixSeq-class object returned by function aa_phychem_index into a 'GRangesMatrixSeq-class' object. Since a 'GRangesMatrixSeq-class' inherits from GRanges-class, this transformation permits the application of several methods from GenomicRanges package in the downstream analysis.

Examples

aln <- c(S1 = "ATGCGGATTAGA", S2 = "ATGACGATCACA", 
        S3 = "ATGAGATCACAG")
cd <- DNAMultipleAlignment(aln)
r1 <- peptide_phychem_index(unmasked(cd), acc = "EISD840101")

r2 <- GRangesMatrixSeq(r1)
r2
#> GRangesMatrixSeq object with 4 ranges and 3 metadata columns:
#>      seqnames    ranges strand |        S1        S2        S3
#>         <Rle> <IRanges>  <Rle> | <numeric> <numeric> <numeric>
#>   A1        1         1      + |      0.26      0.26      0.26
#>   A2        1         2      + |     -1.76     -0.18     -1.76
#>   A3        1         3      + |      0.73      0.73     -0.26
#>   A4        1         4      + |     -1.76     -0.18     -0.69
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome

slot(r2, "phychem")
#> [1] "EISD840101 Consensus normalized hydrophobicity scale (Eisenberg, 1984)"