List of 94 Amino Acid Matrices from AAindex

The aminoacid similarity matrices from Amino Acid Index Database https://www.genome.jp/aaindex/ are provided here. AAindex (ver.9.2) is a database of numerical indices representing various physicochemical and biochemical properties of amino acids and pairs of amino acids.

Usage

data("aaindex2", package = "GenomAutomorphism")

Format

A list carrying the description of 94 Amino Acid Matrices in AAindex ver.9.2 and the text file of matrices imported from https://www.genome.jp/aaindex/.

Details

The similarity of amino acids can be represented numerically, expressed in terms of observed mutation rate or physicochemical properties. A similarity matrix, also called a mutation matrix, is a set of 210 numerical values, 20 diagonal and 20x19/2 off-diagonal elements, used for sequence alignments and similarity searches.

See also

Author

Robersy Sanchez https://genomaths.com

Examples

## Load the mutation matrices from database from the packages
data("aaindex2", package = "GenomAutomorphism")

## Get the available matrices
mat <- aa_mutmat(aaindex = "aaindex2", acc_list = TRUE)
mat[1:10]
#>  [1] "List of 94 Amino Acid Matrices in AAindex ver.9.2"                                
#>  [2] ""                                                                                 
#>  [3] "The columns correspond to the AAindex accession number and the description of"    
#>  [4] "each matrix."                                                                     
#>  [5] ""                                                                                 
#>  [6] "ALTS910101 The PAM-120 matrix (Altschul, 1991)"                                   
#>  [7] "BENS940101 Log-odds scoring matrix collected in 6.4-8.7 PAM (Benner et al., 1994)"
#>  [8] "BENS940102 Log-odds scoring matrix collected in 22-29 PAM (Benner et al., 1994)"  
#>  [9] "BENS940103 Log-odds scoring matrix collected in 74-100 PAM (Benner et al., 1994)" 
#> [10] "BENS940104 Genetic code matrix (Benner et al., 1994)"