Compute the Automorphisms of Mutational Events Between two Codon Sequences Represented in Z64.

Given two codon sequences represented in the Z64 Abelian group, this function computes the automorphisms describing codon mutational events.

autZ64(
  seq = NULL,
  filepath = NULL,
  cube = c("ACGT", "TGCA"),
  cube_alt = c("CATG", "GTAC"),
  start = NA,
  end = NA,
  chr = 1L,
  strand = "+"
)

Arguments

seq
An object from a DNAStringSet or DNAMultipleAlignment class carrying the DNA pairwise alignment of two sequences. The pairwise alignment provided in argument seq or the 'fasta' file filepath must correspond to codon sequences.

filepath
A character vector containing the path to a file in fasta format to be read. This argument must be given if codon & base arguments are not provided.

cube, cube_alt
A character string denoting pairs of the 24 Genetic-code cubes, as given in references (2-3). That is, the base pairs from the given cubes must be complementary each other. Such a cube pair are call dual cubes and, as shown in reference (3), each pair integrates group.

start, end, chr, strand
Optional parameters required to build a GRanges-class. If not provided the default values given for the function definition will be used.

group
A character string denoting the group representation for the given base or codon as shown in reference (1).

Value

An object Automorphism-class with four columns on its metacolumn named: seq1, seq2, autm, and cube.

Details

Automorphisms in Z64 are described as functions \(f(x) = k x mod 64\), where k and x are elements from the set of integers modulo 64.

References

Sanchez R, Morgado E, Grau R. Gene algebra from a genetic code algebraic structure. J Math Biol. 2005 Oct;51(4):431-57. doi: 10.1007/s00285-005-0332-8. Epub 2005 Jul 13. PMID: 16012800. ( PDF).

Robersy Sanchez, Jesús Barreto (2021) Genomic Abelian Finite Groups. doi: 10.1101/2021.06.01.446543.

M. V José, E.R. Morgado, R. Sánchez, T. Govezensky, The 24 possible algebraic representations of the standard genetic code in six or in three dimensions, Adv. Stud. Biol. 4 (2012) 119–152.PDF.

R. Sanchez. Symmetric Group of the Genetic–Code Cubes. Effect of the Genetic–Code Architecture on the Evolutionary Process MATCH Commun. Math. Comput. Chem. 79 (2018) 527-560. PDF.

Examples

## Load a pairwise alignment
data(aln)
aln

#> DNAStringSet object of length 2:
#>     width seq
#> [1]    51 ACCTATGTTGGTATT---GCGCTCCAACTCCTTGGCTCTAGCTCACTACAT
#> [2]    51 ATCTATGTTGGTATTACGACGCTCCAATTCCTTGGGTCC------CTCCTT

## Automorphism on Z64
autms <- autZ64(seq = aln)
autms

#> Automorphism object with 17 ranges and 6 metadata columns:
#>        seqnames    ranges strand |        seq1        seq2    coord1    coord2
#>           <Rle> <IRanges>  <Rle> | <character> <character> <numeric> <numeric>
#>    [1]        1         1      + |         ACC         ATC        17        49
#>    [2]        1         2      + |         TAT         TAT        15        15
#>    [3]        1         3      + |         GTT         GTT        59        59
#>    [4]        1         4      + |         GGT         GGT        43        43
#>    [5]        1         5      + |         ATT         ATT        51        51
#>    ...      ...       ...    ... .         ...         ...       ...       ...
#>   [13]        1        13      + |         TCT         TCC        31        29
#>   [14]        1        14      + |         AGC         ---        52        NA
#>   [15]        1        15      + |         TCA         ---         9        NA
#>   [16]        1        16      - |         CTA         CTC        52        53
#>   [17]        1        17      + |         CAT         CTT         7        55
#>             autm        cube
#>        <numeric> <character>
#>    [1]        33        ACGT
#>    [2]         1        ACGT
#>    [3]         1        ACGT
#>    [4]         1        ACGT
#>    [5]         1        ACGT
#>    ...       ...         ...
#>   [13]         3        ACGT
#>   [14]        NA        None
#>   [15]        NA        None
#>   [16]        30        TGCA
#>   [17]        17        ACGT
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths