Automorphisms between DNA Sequences from two COVID-19 genomes

This is a AutomorphismList object carrying a list of pairwise automorphisms between the SARS coronavirus GZ02 (GenBank: AY390556.1: 265-13398_13398-21485) and Bat SARS-like coronavirus isolate bat-SL-CoVZC45 (GenBank: MG772933.1:265-1345513455-21542), nonstructural_polyprotein. The pairwise DNA sequence alignment is available in the dataset named covid_aln and the automorphisms were estimated with function aut3D.

Usage

data("autm_3d", package = "GenomAutomorphism")

Format

AutomorphismList class object.

Examples

data("autm_3d", package = "GenomAutomorphism")
autm_3d
#> Automorphism object with 9722 ranges and 8 metadata columns:
#>          seqnames    ranges strand |        seq1        seq2         aa1
#>             <Rle> <IRanges>  <Rle> | <character> <character> <character>
#>      [1]        1         1      + |         ATG         ATG           M
#>      [2]        1         2      + |         GAG         GAG           E
#>      [3]        1         3      + |         AGC         AGC           S
#>      [4]        1         4      + |         CTT         CTT           L
#>      [5]        1         5      + |         GTT         GTT           V
#>      ...      ...       ...    ... .         ...         ...         ...
#>   [9718]        1      9718      + |         TCA         TCA           S
#>   [9719]        1      9719      + |         ACT         ACT           T
#>   [9720]        1      9720      + |         CAG         CAG           Q
#>   [9721]        1      9721      + |         GCA         GCA           A
#>   [9722]        1      9722      + |         TAA         TAA           *
#>                  aa2   coord1   coord2        autm        cube
#>          <character> <matrix> <matrix> <character> <character>
#>      [1]           M    1:4:3    1:4:3       1,1,1        ACGT
#>      [2]           E    3:1:3    3:1:3       1,1,1        ACGT
#>      [3]           S    1:3:2    1:3:2       1,1,1        ACGT
#>      [4]           L    2:4:4    2:4:4       1,1,1        ACGT
#>      [5]           V    3:4:4    3:4:4       1,1,1        ACGT
#>      ...         ...      ...      ...         ...         ...
#>   [9718]           S    4:2:1    4:2:1       1,1,1        ACGT
#>   [9719]           T    1:2:4    1:2:4       1,1,1        ACGT
#>   [9720]           Q    2:1:3    2:1:3       1,1,1        ACGT
#>   [9721]           A    3:2:1    3:2:1       1,1,1        ACGT
#>   [9722]           *    4:1:1    4:1:1       1,1,1        ACGT
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths