Autmorphism Grouping by Coefficient

Automorphisms with the same automorphism's coefficients are grouped.

automorphismByCoef(x, ...)

# S4 method for Automorphism
automorphismByCoef(x)

# S4 method for AutomorphismList
automorphismByCoef(
  x,
  min.len = 1L,
  num.cores = detectCores(),
  tasks = 0L,
  verbose = TRUE
)

Arguments

x
An AutomorphismList-class object returned by function automorphism.

num.cores, tasks
Integers. Argument num.cores denotes the number of cores to use, i.e. at most how many child processes will be run simultaneously (see bplapply function from BiocParallel package). Argument tasks denotes the number of tasks per job. value must be a scalar integer >= 0L. In this documentation a job is defined as a single call to a function, such as bplapply. A task is the division of the \(X\) argument into chunks. When tasks == 0 (default), \(X\) is divided as evenly as possible over the number of workers (see MulticoreParam from BiocParallel package).

Value

An AutomorphismByCoef class object. A coefficient with 0 value is assigned to mutational events that are not automorphisms, e.g., indel mutations.

Examples

## Load dataset
data(autm, package = "GenomAutomorphism")

automorphismByCoef(x = autm[1:2])

#> AutomorphismByCoefList object of length 1:
#> $human_1.gorilla
#> AutomorphismByCoef object with 3 ranges and 2 metadata columns:
#>       seqnames    ranges strand |      autm        cube
#>          <Rle> <IRanges>  <Rle> | <numeric> <character>
#>   [1]        1      1-36      + |         1        ACGT
#>   [2]        1        37      + |         3        ACGT
#>   [3]        1    38-106      + |         1        ACGT
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths
#>