Autmorphism Grouping by Coefficient

Automorphisms with the same automorphism's coefficients are grouped.

Usage

automorphism_bycoef(x, ...)

# S4 method for class 'Automorphism'
automorphism_bycoef(x, mut.type = TRUE)

# S4 method for class 'AutomorphismList'
automorphism_bycoef(
  x,
  min.len = 1L,
  mut.type = TRUE,
  num.cores = multicoreWorkers(),
  tasks = 0L,
  verbose = TRUE
)

Arguments

x

An automorphism-class object returned by function automorphisms.

...

Not in use.

mut.type

Logical. Whether to include the mutation type as given by function mut_type.

min.len

Minimum length of a range to be reported.

num.cores, tasks

Integers. Argument num.cores denotes the number of cores to use, i.e. at most how many child processes will be run simultaneously (see bplapply function from BiocParallel package). Argument tasks denotes the number of tasks per job. value must be a scalar integer >= 0L. In this documentation a job is defined as a single call to a function, such as bplapply. A task is the division of the \(X\) argument into chunks. When tasks == 0 (default), \(X\) is divided as evenly as possible over the number of workers (see MulticoreParam from BiocParallel package).

verbose

logic(1). If TRUE, enable progress bar.

Value

An AutomorphismByCoef class object. A coefficient with 0 value is assigned to mutational events that are not automorphisms, e.g., indel mutations.

See also

Examples

## Load dataset
data("autm", package = "GenomAutomorphism")

automorphism_bycoef(x = autm[1:2])
#> AutomorphismByCoef object with 2 ranges and 7 metadata columns:
#>       seqnames    ranges strand |        seq1        seq2         aa1
#>          <Rle> <IRanges>  <Rle> | <character> <character> <character>
#>   [1]        1       1-2      + |         ATG         ATG           M
#>   [2]        1       1-2      + |         GAG         GAG           E
#>               aa2      autm    mut_type        cube
#>       <character> <numeric> <character> <character>
#>   [1]           M         1         HHH        ACGT
#>   [2]           E         1         HHH        ACGT
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths