Codon coordinates on a given a given Abelian group representation.

Given a string denoting a codon or base from the DNA (or RNA) alphabet and a genetic-code Abelian group as given in reference (1).

Usage

codon_coord(codon = NULL, ...)

# S4 method for BaseGroup
codon_coord(codon, group = NULL)

# S4 method for DNAStringSet_OR_NULL
codon_coord(
  codon = NULL,
  filepath = NULL,
  cube = c("ACGT", "AGCT", "TCGA", "TGCA", "CATG", "GTAC", "CTAG", "GATC", "ACTG",
    "ATCG", "GTCA", "GCTA", "CAGT", "TAGC", "TGAC", "CGAT", "AGTC", "ATGC", "CGTA",
    "CTGA", "GACT", "GCAT", "TACG", "TCAG"),
  group = c("Z4", "Z5", "Z64", "Z125", "Z4^3", "Z5^3"),
  start = NA,
  end = NA,
  chr = 1L,
  strand = "+"
)

# S4 method for matrix_OR_data_frame
codon_coord(
  codon,
  cube = c("ACGT", "AGCT", "TCGA", "TGCA", "CATG", "GTAC", "CTAG", "GATC", "ACTG",
    "ATCG", "GTCA", "GCTA", "CAGT", "TAGC", "TGAC", "CGAT", "AGTC", "ATGC", "CGTA",
    "CTGA", "GACT", "GCAT", "TACG", "TCAG"),
  group = c("Z64", "Z125", "Z4^3", "Z5^3")
)

Arguments

codon

An object from BaseGroup-class (generated with function base_coord), DNAStringSet or from DNAMultipleAlignment class carrying the DNA pairwise alignment of two sequences.

...

Not in use.

group

A character string denoting the group representation for the given base or codon as shown in reference (2-3).

filepath

A character vector containing the path to a file in fasta format to be read. This argument must be given if codon & base arguments are not provided.

cube

A character string denoting one of the 24 Genetic-code cubes, as given in references (2-3).

start, end, chr, strand

Optional parameters required to build a GRanges-class. If not provided the default values given for the function definition will be used.

Value

A CodonGroup-class object.

Details

Symbols "-" and "N" usually found in DNA sequence alignments to denote gaps and missing/unknown bases are represented by the number: '-1' on Z4 and '0' in Z5. In Z64 the symbol 'NA' will be returned for codons including symbols "-" and "N".

This function returns a GRanges-class object carrying the codon sequence(s) and their respective coordinates in the requested Abelian group or simply, when group = 'Z5^3' 3D-coordinates, which are derive from Z5 as indicated in reference (3). Notice that the coordinates can be 3D or just one-dimension ("Z64" or "Z125"). Hence, the pairwise alignment provided in argument codon must correspond to codon sequences.

References

Robersy Sanchez, Jesús Barreto (2021) Genomic Abelian Finite Groups. doi: 10.1101/2021.06.01.446543.

M. V José, E.R. Morgado, R. Sánchez, T. Govezensky, The 24 possible algebraic representations of the standard genetic code in six or in three dimensions, Adv. Stud. Biol. 4 (2012) 119–152.PDF.

R. Sanchez. Symmetric Group of the Genetic–Code Cubes. Effect of the Genetic–Code Architecture on the Evolutionary Process MATCH Commun. Math. Comput. Chem. 79 (2018) 527-560.

See also

Author

Robersy Sanchez https://genomaths.com

Examples

## Load a pairwise alignment
data(aln)
aln
#> DNAStringSet object of length 2:
#>     width seq
#> [1]    51 ACCTATGTTGGTATT---GCGCTCCAACTCCTTGGCTCTAGCTCACTACAT
#> [2]    51 ATCTATGTTGGTATTACGACGCTCCAATTCCTTGGGTCC------CTCCTT

## DNA base representation in the Abelian group Z5
bs_cor <- codon_coord(
    codon = aln,
    cube = "ACGT",
    group = "Z5"
)
bs_cor ## 3-D coordinates
#> CodonGroup object with 17 ranges and 4 metadata columns:
#>        seqnames    ranges strand |        seq1        seq2      coord1
#>           <Rle> <IRanges>  <Rle> | <character> <character> <character>
#>    [1]        1         1      + |         A-G         AGG       1,0,3
#>    [2]        1         2      + |         CGC         TAG       2,3,2
#>    [3]        1         3      + |         CCT         CCT       2,2,4
#>    [4]        1         4      + |         TGC         TGC       4,3,2
#>    [5]        1         5      + |         ACT         ACC       1,2,4
#>    ...      ...       ...    ... .         ...         ...         ...
#>   [13]        1        13      + |         ACT         ACT       1,2,4
#>   [14]        1        14      + |         TCA         TCC       4,2,1
#>   [15]        1        15      + |         TTC         TTC       4,4,2
#>   [16]        1        16      + |         -TA         ATT       0,4,1
#>   [17]        1        17      + |         -GT         CGT       0,3,4
#>             coord2
#>        <character>
#>    [1]       1,3,3
#>    [2]       4,1,3
#>    [3]       2,2,4
#>    [4]       4,3,2
#>    [5]       1,2,2
#>    ...         ...
#>   [13]       1,2,4
#>   [14]       4,2,2
#>   [15]       4,4,2
#>   [16]       1,4,4
#>   [17]       2,3,4
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths

## DNA base representation in the Abelian group Z64
bs_cor <- codon_coord(
    codon = aln,
    cube = "ACGT",
    group = "Z64"
)
bs_cor
#> CodonGroup object with 17 ranges and 4 metadata columns:
#>        seqnames    ranges strand |        seq1        seq2      coord1
#>           <Rle> <IRanges>  <Rle> | <character> <character> <character>
#>    [1]        1         1      + |         A-G         AGG        <NA>
#>    [2]        1         2      + |         CGC         TAG          37
#>    [3]        1         3      + |         CCT         CCT          23
#>    [4]        1         4      + |         TGC         TGC          45
#>    [5]        1         5      + |         ACT         ACC          19
#>    ...      ...       ...    ... .         ...         ...         ...
#>   [13]        1        13      + |         ACT         ACT          19
#>   [14]        1        14      + |         TCA         TCC          28
#>   [15]        1        15      + |         TTC         TTC          61
#>   [16]        1        16      + |         -TA         ATT        <NA>
#>   [17]        1        17      + |         -GT         CGT        <NA>
#>             coord2
#>        <character>
#>    [1]          34
#>    [2]          14
#>    [3]          23
#>    [4]          45
#>    [5]          17
#>    ...         ...
#>   [13]          19
#>   [14]          29
#>   [15]          61
#>   [16]          51
#>   [17]          39
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths

## Giving a matrix of codons
codon_coord(base2codon(x = aln))
#>    seq1 seq2 coord1 coord2 NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA
#> 1   ACC  ATC     17     49 15 15 59 59 43 43 51 51 NA 18 26 18 53 53  4  4 53
#> 2   TAT  TAT     17     49 15 15 59 59 43 43 51 51 NA 18 26 18 53 53  4  4 53
#> 3   GTT  GTT     17     49 15 15 59 59 43 43 51 51 NA 18 26 18 53 53  4  4 53
#> 4   GGT  GGT     17     49 15 15 59 59 43 43 51 51 NA 18 26 18 53 53  4  4 53
#> 5   ATT  ATT     17     49 15 15 59 59 43 43 51 51 NA 18 26 18 53 53  4  4 53
#> 6   ---  ACG     17     49 15 15 59 59 43 43 51 51 NA 18 26 18 53 53  4  4 53
#> 7   GCG  ACG     17     49 15 15 59 59 43 43 51 51 NA 18 26 18 53 53  4  4 53
#> 8   CTC  CTC     17     49 15 15 59 59 43 43 51 51 NA 18 26 18 53 53  4  4 53
#> 9   CAA  CAA     17     49 15 15 59 59 43 43 51 51 NA 18 26 18 53 53  4  4 53
#> 10  CTC  TTC     17     49 15 15 59 59 43 43 51 51 NA 18 26 18 53 53  4  4 53
#> 11  CTT  CTT     17     49 15 15 59 59 43 43 51 51 NA 18 26 18 53 53  4  4 53
#> 12  GGC  GGG     17     49 15 15 59 59 43 43 51 51 NA 18 26 18 53 53  4  4 53
#> 13  TCT  TCC     17     49 15 15 59 59 43 43 51 51 NA 18 26 18 53 53  4  4 53
#> 14  AGC  ---     17     49 15 15 59 59 43 43 51 51 NA 18 26 18 53 53  4  4 53
#> 15  TCA  ---     17     49 15 15 59 59 43 43 51 51 NA 18 26 18 53 53  4  4 53
#> 16  CTA  CTC     17     49 15 15 59 59 43 43 51 51 NA 18 26 18 53 53  4  4 53
#> 17  CAT  CTT     17     49 15 15 59 59 43 43 51 51 NA 18 26 18 53 53  4  4 53
#>    NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA
#> 1  61 55 55 41 42 31 29 33 NA 28 NA 52 53  7 55
#> 2  61 55 55 41 42 31 29 33 NA 28 NA 52 53  7 55
#> 3  61 55 55 41 42 31 29 33 NA 28 NA 52 53  7 55
#> 4  61 55 55 41 42 31 29 33 NA 28 NA 52 53  7 55
#> 5  61 55 55 41 42 31 29 33 NA 28 NA 52 53  7 55
#> 6  61 55 55 41 42 31 29 33 NA 28 NA 52 53  7 55
#> 7  61 55 55 41 42 31 29 33 NA 28 NA 52 53  7 55
#> 8  61 55 55 41 42 31 29 33 NA 28 NA 52 53  7 55
#> 9  61 55 55 41 42 31 29 33 NA 28 NA 52 53  7 55
#> 10 61 55 55 41 42 31 29 33 NA 28 NA 52 53  7 55
#> 11 61 55 55 41 42 31 29 33 NA 28 NA 52 53  7 55
#> 12 61 55 55 41 42 31 29 33 NA 28 NA 52 53  7 55
#> 13 61 55 55 41 42 31 29 33 NA 28 NA 52 53  7 55
#> 14 61 55 55 41 42 31 29 33 NA 28 NA 52 53  7 55
#> 15 61 55 55 41 42 31 29 33 NA 28 NA 52 53  7 55
#> 16 61 55 55 41 42 31 29 33 NA 28 NA 52 53  7 55
#> 17 61 55 55 41 42 31 29 33 NA 28 NA 52 53  7 55