Automorphisms between DNA Sequences from two COVID-19 genomes

This is a AutomorphismList object carrying a list of pairwise automorphisms between the SARS coronavirus GZ02 (GenBank: AY390556.1: 265-13398_13398-21485) and Bat SARS-like coronavirus isolate bat-SL-CoVZC45 (GenBank: KY417151.1: protein-coding regions). The pairwise DNA sequence alignment is available in the dataset named covid_aln and the automorphisms were estimated with function autZ64.

Usage

data("covid_autm", package = "GenomAutomorphism")

Format

AutomorphismList class object.

See also

Author

Robersy Sanchez https://genomaths.com

Examples

data("covid_autm", package = "GenomAutomorphism",
    envir = environment())
covid_autm
#> Automorphism object with 9722 ranges and 8 metadata columns:
#>          seqnames    ranges strand |        seq1        seq2         aa1
#>             <Rle> <IRanges>  <Rle> | <character> <character> <character>
#>      [1]        1         1      + |         ATG         ATG           M
#>      [2]        1         2      + |         GAG         GAG           E
#>      [3]        1         3      + |         AGC         AGC           S
#>      [4]        1         4      + |         CTT         CTT           L
#>      [5]        1         5      + |         GTT         GTT           V
#>      ...      ...       ...    ... .         ...         ...         ...
#>   [9718]        1      9718      + |         TCA         TCA           S
#>   [9719]        1      9719      + |         ACT         ACT           T
#>   [9720]        1      9720      + |         CAG         CAG           Q
#>   [9721]        1      9721      + |         GCA         GCA           A
#>   [9722]        1      9722      + |         TAA         TAA           *
#>                  aa2    coord1    coord2      autm        cube
#>          <character> <numeric> <numeric> <numeric> <character>
#>      [1]           M        50        50         1        ACGT
#>      [2]           E        10        10         1        ACGT
#>      [3]           S        33        33         1        ACGT
#>      [4]           L        55        55         1        ACGT
#>      [5]           V        59        59         1        ACGT
#>      ...         ...       ...       ...       ...         ...
#>   [9718]           S        28        28         1        ACGT
#>   [9719]           T        19        19         1        ACGT
#>   [9720]           Q         6         6         1        ACGT
#>   [9721]           A        24        24         1        ACGT
#>   [9722]           *        12        12         1        ACGT
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths