Multiple Sequence Alignment (MSA) of Primate Somatic Cytochrome C

This is a DNAMultipleAlignment carrying a MSA of Primate Somatic Cytochrome C to be used in the examples provided for the package functions. The original file can be downloaded from GitHub at: https://bit.ly/3kdEAzs

Usage

data("cyc_aln", package = "GenomAutomorphism")

Format

DNAMultipleAlignment class object.

See also

Author

Robersy Sanchez https://genomaths.com

Examples

data("cyc_aln", package = "GenomAutomorphism")
cyc_aln
#> DNAMultipleAlignment with 19 rows and 318 columns
#>       aln                                                   names               
#>  [1] ATGGGTGATGTTGAGAAAGGCAAGA...TCTCAAAAAAGCTACTAATGAGTAA NM_018947.6 Homo ...
#>  [2] ATGGGTGATGTTGAGAAAGGCAAGA...TCTCAAAAAAGCTACTAATGAGTAA gb|BC021994.1|  [...
#>  [3] ATGGGTGATGTTGAGAAAGGCAAGA...TCTCAAAAAAGCTACTAATGAGTAA ref|XM_019031021....
#>  [4] ATGGGTGATGTTGAGAAAGGCAAGA...TCTCAAAAAAGCTACTAATGAGTAA dbj|AK311836.1|  ...
#>  [5] ATGGGTGATGTTGAGAAAGGCAAGA...TCTCAAAAAAGCTACTAATGAGTAA gb|BC067222.1|  [...
#>  [6] ATGGGTGATGTTGAGAAAGGCAAGA...TCTCAAAAAAGCTACTAATGAGTAA gb|BC068464.1|  [...
#>  [7] ATGGGTGATGTTGAGAAAGGCAAGA...TCTCAAAAAAGCTACTAATGAGTAA gb|BC015130.1|  [...
#>  [8] ATGGGTGATGTTGAGAAAGGCAAGA...TCTCAAAAAAGCTACTAATGAGTAA ref|XM_032759077....
#>  [9] ATGGGTGATGTTGAGAAAGGCAAGA...TCTCAAAAAAGCTACTAATGAGTAA ref|XM_003270423....
#> [10] ATGGGTGATGTTGAGAAAGGCAAGA...TCTCAAAAAAGCTACTAATGAGTAA ref|XM_033192823....
#> [11] ATGGGTGATGTTGAGAAAGGCAAGA...TCTCAAAAAAGCTACTAATGAGTAA ref|XM_003896182....
#> [12] ATGGGTGATGTTGAGAAAGGCAAGA...TCTCAAAAAAGCTACTAATGAGTAA ref|XM_003895240....
#> [13] ATGGGTGATGTTGAGAAAGGCAAGA...TCTCAAAAAAGCTACTAATGAGTAA ref|XM_010375926....
#> [14] ATGGGTGATGTTGAGAAAGGCAAGA...TCTCAAAAAAGCTACTAATGAGTAA ref|XM_028845949....
#> [15] ATGGGTGATGTTGAGAAAGGCAAGA...TCTCAAAAAAGCTACTAATGAGTAA ref|XM_028845948....
#> [16] ATGGGTGATGTTGAGAAAGGCAAGA...TCTCAAAAAAGCTACTAATGAGTAA ref|XM_025380676....
#> [17] ATGGGTGATGTTGAGAAAGGCAAGA...TCTCAAAAAAGCTACTAATGAGTAA ref|XM_025380675....
#> [18] ATGGGTGATGTTGAGAAAGGCAAGA...TCTCAAAAAAGCTACTAATGAGTAA ref|NM_001131167....
#> [19] ATGGGTGATGTTGAGAAAGGCAAGA...TCTCAAAAAAGCTACTAATGAGTAA emb|CR857183.1|  ...