Automorphisms between DNA Sequences from Primate Cytochrome C Genes

This is a AutomorphismList object carrying a list of pairwise automorphisms between the DNA sequences from the MSA of Primate Somatic Cytochrome C to be used in the examples provided for the package functions. The automorphisms were estimated from the cyc_aln MSA with function autZ64.

Usage

data("cyc_autm", package = "GenomAutomorphism")

Format

AutomorphismList class object.

See also

Author

Robersy Sanchez https://genomaths.com

Examples

data("cyc_autm", package = "GenomAutomorphism")
cyc_autm
#> AutomorphismList object of length: 171
#> names(171): human_1.human_2 human_1.gorilla human_1.human_3 ... gelada_baboon_2.orangutan_1 gelada_baboon_2.orangutan_2 orangutan_1.orangutan_2 
#> ------- 
#> Automorphism object with 106 ranges and 8 metadata columns:
#>         seqnames    ranges strand |        seq1        seq2         aa1
#>            <Rle> <IRanges>  <Rle> | <character> <character> <character>
#>     [1]        1         1      + |         ATG         ATG           M
#>     [2]        1         2      + |         GGT         GGT           G
#>     [3]        1         3      + |         GAT         GAT           D
#>     [4]        1         4      + |         GTT         GTT           V
#>     [5]        1         5      + |         GAG         GAG           E
#>     ...      ...       ...    ... .         ...         ...         ...
#>   [102]        1       102      + |         GCT         GCT           A
#>   [103]        1       103      + |         ACT         ACT           T
#>   [104]        1       104      + |         AAT         AAT           N
#>   [105]        1       105      + |         GAG         GAG           E
#>   [106]        1       106      + |         TAA         TAA           *
#>                 aa2    coord1    coord2      autm        cube
#>         <character> <numeric> <numeric> <numeric> <character>
#>     [1]           M        50        50         1        ACGT
#>     [2]           G        43        43         1        ACGT
#>     [3]           D        11        11         1        ACGT
#>     [4]           V        59        59         1        ACGT
#>     [5]           E        10        10         1        ACGT
#>     ...         ...       ...       ...       ...         ...
#>   [102]           A        27        27         1        ACGT
#>   [103]           T        19        19         1        ACGT
#>   [104]           N         3         3         1        ACGT
#>   [105]           E        10        10         1        ACGT
#>   [106]           *        12        12         1        ACGT
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths
#> ...
#> <170 more DFrame element(s)>
#> Two slots: 'DataList' & 'SeqRanges'
#> -------