Classification of DNA base mutations

Each DNA/RNA base can be classified into three main classes according to three criteria (1): number of hydrogen bonds (strong-weak), chemical type (purine-pyrimidine), and chemical groups (amino versus keto). Each criterion produces a partition of the set of bases: 1. According to the number of hydrogen bonds (on DNA/RNA double helix): strong \(S=(C,G)\) (three hydrogen bonds) and weak \(W=(A,U)\) (two hydrogen bonds); 2. According to the chemical type: purines \(R=(A, G)\) and pyrimidines \(Y=(C,U)\). 3. According to the presence of amino or keto groups on the base rings: amino \(M=(C,A)\) and keto \(K=(G,U)\). So, each mutational event can be classified as according to the type of involved in it (2).

Usage

mut_type(x, y)

Arguments

x, y

Character strings denoting DNA bases

Value

A character string of same length of 'x' and 'y'.

References

A. Cornish-Bowden, Nomenclature for incompletely specified bases in nucleic acid sequences: recommendations 1984, Nucleic Acids Res. 13 (1985) 3021-3030.

MA.A. Jimenez-Montano, C.R. de la Mora-Basanez, T. Poschel, The hypercube structure of the genetic code explains conservative and non-conservative aminoacid substitutions in vivo and in vitro, Biosystems. 39 (1996) 117-125.

Examples

## Mutation type 'R'
mut_type("A", "G")
#> [1] "R"

## Mutation type 'M'
mut_type("A", "C")
#> [1] "M"

## Mutation type 'W'
mut_type("A", "T")
#> [1] "W"

## Mutation type 'S'
mut_type("G", "C")
#> [1] "S"