Autmorphism Grouping by Coefficient

Automorphisms with the same automorphism's coefficients are grouped.

Usage

automorphismByCoef(x, ...)

# S4 method for Automorphism
automorphismByCoef(x, mut.type = TRUE)

# S4 method for AutomorphismList
automorphismByCoef(
  x,
  min.len = 1L,
  mut.type = TRUE,
  num.cores = detectCores() - 1,
  tasks = 0L,
  verbose = TRUE
)

Arguments

x

An automorphism-class object returned by function automorphism.

mut.type

Logical. Whether to include the mutation type as given by function mut_type.

num.cores, tasks

Integers. Argument num.cores denotes the number of cores to use, i.e. at most how many child processes will be run simultaneously (see bplapply function from BiocParallel package). Argument tasks denotes the number of tasks per job. value must be a scalar integer >= 0L. In this documentation a job is defined as a single call to a function, such as bplapply. A task is the division of the \(X\) argument into chunks. When tasks == 0 (default), \(X\) is divided as evenly as possible over the number of workers (see MulticoreParam from BiocParallel package).

Value

An AutomorphismByCoef class object. A coefficient with 0 value is assigned to mutational events that are not automorphisms, e.g., indel mutations.

Examples

## Load dataset
data(autm, package = "GenomAutomorphism")

automorphismByCoef(x = autm[1:2])
#> AutomorphismByCoef object with 2 ranges and 5 metadata columns:
#>       seqnames    ranges strand |        seq1        seq2      autm    mut_type
#>          <Rle> <IRanges>  <Rle> | <character> <character> <numeric> <character>
#>   [1]        1       1-2      + |         ATG         ATG         1         HHH
#>   [2]        1       1-2      + |         GAG         GAG         1         HHH
#>              cube
#>       <character>
#>   [1]        ACGT
#>   [2]        ACGT
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths