DNA numerical matrix

This function applies the numerical indices representing various physicochemical and biochemical properties of DNA bases. As results, DNA sequences are represented as numerical vectors which can be subject of further downstream statistical analysis and digital signal processing.

Usage

dna_phychem(seqs, ...)

# S4 method for class 'character'
dna_phychem(
  seqs,
  phychem = list(A = NULL, T = NULL, C = NULL, G = NULL, N = NULL)
)

# S4 method for class 'DNAStringSet_OR_DNAMultipleAlignment'
dna_phychem(
  seqs,
  phychem = list(A = NULL, T = NULL, C = NULL, G = NULL, N = NULL),
  index_name = NULL,
  ...
)

Arguments

seqs

A character string, a DNAStringSet or a DNAMultipleAlignment class object carrying the DNA pairwise alignment of two sequences.

...

Not in use.

phychem

A list of DNA bases physicochemical properties, e.g., like those provided in dna_phyche.

index_name

Optional. Name of breve description of the base physicochemical property applied to represent the DNA sequence.

Value

A MatrixSeq-class object.

See also

Author

Robersy Sanchez https://genomaths.com

Examples

## Let's create DNAStringSet-class object 
base <- DNAStringSet(x = c( seq1 ='ACGTGATCAAGT', 
                            seq2 = 'GTGTGATCCAGT', 
                            seq3 = 'TCCTGATCAGGT'))

dna_phychem(seqs = base,
            phychem = list('A' = 0.87, 'C' = 0.88, 'T' = 0.82,
                            'G' = 0.89, 'N' = NA),
            index_name = "Proton-Affinity")
#> MatrixSeq with 3 rows (sequences) and 12 columns (aminoacids/codons):
#> ------- 
#>      B1   B2   B3   B4   B5   B6   B7   B8   B9  B10  B11  B12
#> S1 0.87 0.88 0.89 0.82 0.89 0.87 0.82 0.88 0.87 0.87 0.89 0.82
#> S2 0.89 0.82 0.89 0.82 0.89 0.87 0.82 0.88 0.88 0.87 0.89 0.82
#> S3 0.82 0.88 0.88 0.82 0.89 0.87 0.82 0.88 0.87 0.89 0.89 0.82
#> ------- 
#> Slots: 'seqs', 'matrix', 'names', 'aaindex', 'phychem', 'accession