Get Mutation Score from an AAindex Matrix

This function is applied to get the mutation or contact potential scores representing the similarity/distance between amino acids corresponding to substitution mutations. The score are retrieve from a mutation matrix or a statistical protein contact potentials matrix from AAindex (ver.9.2).

Usage

get_mutscore(aa1, aa2, ...)

# S4 method for character,character
get_mutscore(
  aa1,
  aa2,
  acc = NULL,
  aaindex = NULL,
  mutmat = NULL,
  alphabet = c("AA", "DNA"),
  num.cores = 1L,
  tasks = 0L,
  verbose = FALSE,
  ...
)

Arguments

aa1, aa2

A simple character representing an amino acids or a character string of letter from the amino acid alphabet or base-triplets from the DNA/RNA alphabet.

...

Not in use.

acc

Accession id for a specified mutation or contact potential matrix.

aaindex

Database where the requested accession id is locate. The possible values are: "aaindex2" or "aaindex3".

mutmat

A mutation or any score matrix provided by the user.

alphabet

Whether the alphabet is from the 20 amino acid (AA) or four (DNA)/RNA base alphabet. This would prevent mistakes, i.e., the strings "ACG" would be a base-triplet on the DNA alphabet or simply the amino acid sequence of alanine, cysteine, and glutamic acid.

num.cores, tasks

Parameters for parallel computation using package BiocParallel-package: the number of cores to use, i.e. at most how many child processes will be run simultaneously (see bplapply and the number of tasks per job (only for Linux OS).

verbose

If TRUE, prints the function log to stdout.

Value

A single numeric score or a numerical vector.

Details

If a score matrix is provided by the user, then it must be a symmetric matrix 20x20.

See also

Author

Robersy Sanchez https://genomaths.com

Examples

## Load the mutation matrices from database from the packages
data("aaindex2", package = "GenomAutomorphism" )

## A single amino acids substitution mutation
get_mutscore("A", "C", acc = "MIYS930101", aaindex = aaindex2)
#> [1] -0.51

## A tri-peptide mutation
get_mutscore(aa1 = "ACG", aa2 = "ATG", acc = "MIYS930101",
            aaindex = aaindex2, alphabet = "AA")
#> [1]  0.34 -0.62  0.43

## A single base-triple mutation, i.e., a single amino acid substitution 
## mutation
get_mutscore(aa1 = "ACG", aa2 = "CTA", acc = "MIYS930101",
            aaindex = aaindex2, alphabet = "DNA")
#> [1] -0.77

## Peptides can be also written as:
get_mutscore(aa1 = c("A","C","G"), aa2 = c("C","T","A"), 
            acc = "MIYS930101", aaindex = aaindex2, alphabet = "AA")
#> [1] -0.51 -0.62  0.19